Add packer build for docker container
* Add docker Ubuntu Trusty packer template.
* Rework cmtools to install ansible and puppet
* Add support for Vagrant provider docker.
* Containers will share the host system docker daemon
  and images in the storage, so may be run nested.
* Ensure the rsyslogd and sshd is running in containers.
* Also ensure the /var/log/solar dir, which is required for docker.
* W/a Solar-hardcoded and vbox specific ssh key paths
  for docker case.
* Update FAQ and docs

Depends-on: I881d362968a10d816cbd368cb185900dfaa0b3bc
Partial-bug: #1547587

Note, vagrant does not support vm.network and exec based
provisioning for docker, so there are workarounds.

+++ bootstrap/playbooks/files/cleanup.sh
@@ -48,21 +48,23 @@ echo "==> Clearing last login information"
 >/var/log/wtmp
 >/var/log/btmp
 
-# Whiteout root
-count=$(df --sync -kP / | tail -n1  | awk -F ' ' '{print $4}')
-let count--
-dd if=/dev/zero of=/tmp/whitespace bs=1024 count=$count
-rm /tmp/whitespace
+if [ "${cleanup}" = "true" ] ; then
+    # Whiteout root
+    count=$(df --sync -kP / | tail -n1  | awk -F ' ' '{print $4}')
+    let count--
+    dd if=/dev/zero of=/tmp/whitespace bs=1024 count=$count
+    rm /tmp/whitespace
 
-# Whiteout /boot
-count=$(df --sync -kP /boot | tail -n1 | awk -F ' ' '{print $4}')
-let count--
-dd if=/dev/zero of=/boot/whitespace bs=1024 count=$count
-rm /boot/whitespace
+    # Whiteout /boot
+    count=$(df --sync -kP /boot | tail -n1 | awk -F ' ' '{print $4}')
+    let count--
+    dd if=/dev/zero of=/boot/whitespace bs=1024 count=$count
+    rm /boot/whitespace
 
-# Zero out the free space to save space in the final image
-dd if=/dev/zero of=/EMPTY bs=1M
-rm -f /EMPTY
+    # Zero out the free space to save space in the final image
+    dd if=/dev/zero of=/EMPTY bs=1M
+    rm -f /EMPTY
+fi
 
 # Make sure we wait until all the data is written to disk, otherwise
 # Packer might quite too early before the large files are deleted
